Add files via upload
That's a first more extense program that I write. Got the idea from an exercise proposed in my "Learn Python 3" course via Codecademy. Don't know if it is any good. Probably could have done it a lot simpler with a little bit more knowledge. But, found out to be interesting what I could do with the little that I know. Hope it's not just my beginners perception, although I know it most certainly is hahaha. It has been a lot of fun to explore Python.